+++ Cpp09/ex01/main.cpp
@@ -13,7 +13,7 @@ int main(int argc, char **argv)
 
 	try
 	{
-		int result = rpn.evaluate(argv[1]);
+		long long result = rpn.evaluate(argv[1]);
 		std::cout << result << std::endl;
 	}
 	catch (const std::exception &e)
